double

+++ exercises/02_arrays/reverse.cc
@@ -29,7 +29,7 @@ int main(int argc, char* argv[]) {
     std::cin>>len;
 
     
-    int* v= init_vector<int>(len);
+    double* v= init_vector<int>(len);
  
     for (int i=0; i < len; ++i)
     std::cin >> v[i];
